Blackjack 클래스 생성 후 migration
